+++ src/refactoring/hooks/utils/productUtils.ts
@@ -7,3 +7,6 @@ export const getRemainingStock = (product: Product, cartItem?: CartItem) => {
 export const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
   return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
 };
+
+export const getProductById = (products: Product[], productId: string) =>
+  products.find((p) => p.id === productId);
